renamed folder source to models
